scripts/ruuvitag-manager/requirements.txt
paho-mqtt>=1.4.0
ruuvitag_sensor>=0.11.0